backend/requirements.txt
alembic==1.16.5
Flask==3.0.3
gunicorn==22.0.0
opentelemetry-api>=1.24,<2
opentelemetry-exporter-otlp-proto-http>=1.24,<2
opentelemetry-instrumentation-flask>=0.45b0,<1
opentelemetry-instrumentation-sqlalchemy>=0.45b0,<1
opentelemetry-sdk>=1.24,<2
prometheus-client>=0.20,<1
pydantic>=2,<3
psycopg[binary]==3.2.6
redis>=5,<6
SQLAlchemy==2.0.40
strawberry-graphql==0.314.3
